Rename boolean helper file
